Move action constants to action creator files.
