+++ src/actions/index.js
@@ -1,3 +1,2 @@
-export * from './types'
 export * from './securities'
 export * from './counter'
