--- src/actions/types.js
@@ -1,7 +1,0 @@
-export const SECURITIES_FETCH = 'SECURITIES_FETCH'
-export const SECURITIES_FETCH_SUCCESS = 'SECURITIES_FETCH_SUCCESS'
-export const SECURITIES_FETCH_FAILURE = 'SECURITIES_FETCH_FAILURE'
-
-export const COUNTER_INCREMENT = 'COUNTER_INCREMENT'
-export const COUNTER_DECREMENT = 'COUNTER_DECREMENT'
-export const COUNTER_SET_SIZE = 'COUNTER_SET_SIZE'
