+++ src/reducers/securities.js
@@ -1,4 +1,4 @@
-import { SECURITIES_FETCH, SECURITIES_FETCH_SUCCESS, SECURITIES_FETCH_FAILURE } from '../actions//types'
+import { SECURITIES_FETCH, SECURITIES_FETCH_SUCCESS, SECURITIES_FETCH_FAILURE } from '../actions'
 
 const initialState = {}
